chore: move cards logic to cards widget
